@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -26,10 +26,10 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  sqflite: ^1.3.1
-  uuid: ^2.1.0
-  provider: ^4.1.3
+  cupertino_icons: ^1.0.3
+  sqflite: ^2.0.0+3
+  uuid: ^3.0.4
+  provider: ^5.0.0
 
 dev_dependencies:
   flutter_test:
